New API for Button handling
Allows you to easily configure multiple actions for one button due to different methods of pressing it.
This allows you to unload the number of buttons involved, which are often in short supply. For example, you can assign the actions of switching the range of your voice to a double or triple press of the same button on which the microphone is bound (N by default). To avoid false triggering, there is a method `OnTotalRepeatedPress` and `CallbackOnLongPressDetection` that accurately assess the intention of pressing.